Refactor HeatTreatmentChart and move it to new location
The HeatTreatmentChart component has been refactored and moved from the packages/charts directory to the apps/web components directory. The data used in the chart has been modified, as well as the dimensions of the chart and the color scheme. Code cleaning has also been done to remove unnecessary lines and improve readability.